Watch directories; always go into user-specified directories
When listing files to wait on, we currently never go into a directory
starting with '.', even if the user explicitly asks us to wait on '.git/'.
We change this to always look one level inside the given directory, and to
recurse only into non-hidden directories.

Also, explicitly wait on directories, to pick up new files, since a
directory's modification time changes when a file is added, so want to
watch both the files (for changes) and the directories (for new files).