Handle GHCI unexpected errors better
First reported in
https://github.com/obsidiansystems/obelisk/issues/563, we see an issue
where ghci crashes had bad error messages in ghcid. Only information
given was "unexpected error" and the command that was run.

To better handle this, we need to modify the “consume” functions to
better keep track of error messages. This changes the type signature
from a Maybe to an Either:

  -consume :: Stream -> (String -> IO (Maybe a)) -> IO (Maybe a)
  +consume :: Stream -> (String -> IO (Maybe a)) -> IO (Either (Maybe String) a)

This allows us to keep track of what the last message from GHCI was so
so that we can report it back to the user. Similarly, the signature
for GhciError was changed:

  -data GhciError = UnexpectedExit String String
  +data GhciError = UnexpectedExit String String (Maybe String)

So that we can store this information. After that, just change the
error handling to print a Just "..." message.

ghcid now shows something like this on out of memory:

  Command "ghci ..." exited unexpectedly with error message: ghc:
  failed to create OS thread: Cannot allocate memory