Report GHC panics
Here is an example of a GHC panic that `ghcid` was hiding:

```
ghc: panic! (the 'impossible' happened)
  (GHC version 8.6.5 for x86_64-unknown-linux):
        buildKindCoercion
  Div ((rp_a1iFz * (1 + (64 * (CLog 2 64 + 8)))) - 1) 18
  Div ((1 + (64 * (CLog 2 64 + 8))) - 1) WeightRamWidth
  CLog 2 64 + 8
  8
  Call stack:
      CallStack (from HasCallStack):
        callStackDoc, called at compiler/utils/Outputable.hs:1160:37 in ghc:Outputable
        pprPanic, called at compiler/types/Coercion.hs:2069:9 in ghc:Coercion

Please report this as a GHC bug:  http://www.haskell.org/ghc/reportabug
```

This commit parses this type of error in order for `ghcid` to report
them properly.

Note: the last line is not kept as part of the error message. Maybe it
should be, but it is not obvious to me how to do so since it doesn't
seem to follow the same leading whitespace format as the rest of the
error message.