Ignore spurious message in ghcid.txt
When ghcid is exited, it writes "Ghcid has stopped." to the
log file. The VS Code extension tries to parse that as an error,
but there is no file to open and no line number to show. VS Code isn't
smart enough to ignore the error; it just hangs around forever (or until
you delete the log file).

This PR adds that line as one to ignore.